auth-service/auth_requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
cognitive-complexity==1.3.0
colorama==0.4.6
distlib==0.3.7
dnspython==2.4.2
email-validator==2.0.0
eradicate==2.3.0
exceptiongroup==1.1.3
fastapi==0.100.1
filelock==3.12.3
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
identify==2.5.27
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mr-proper==0.0.7
multidict==6.0.4
nodeenv==1.8.0
orjson==3.9.2
packaging==23.1
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.3.3
psycopg==3.1.10
pycodestyle==2.10.0
pydantic==2.1.1
pydantic-settings==2.0.1
pydantic_core==2.4.0
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.4.1
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
stdlib-list==0.9.0
tenacity==8.2.2
tomli==2.0.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
virtualenv==20.24.5
Werkzeug==2.3.7
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
